% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM.R
\name{balloonGgPlot}
\alias{balloonGgPlot}
\title{balloonGgPlot}
\usage{
balloonGgPlot(M, nscale, logTran, sqrtTran, xlabel, ylabel, main)
}
\arguments{
\item{M}{A matrix}

\item{nscale}{scaling number}

\item{logTran}{Whether to do log-transformation}

\item{sqrtTran}{Whether to do sqrt-transformation}

\item{xlabel}{label for rows}

\item{ylabel}{label for columns}

\item{main}{title}
}
\value{
A balloon plot for the matrix
}
\description{
balloonGgPlot
}
\keyword{BalloonPlot}
\keyword{ggplot}
