% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeCon.R
\name{center_scale}
\alias{center_scale}
\title{center_scale}
\usage{
center_scale(Mat, center = FALSE, scale = TRUE)
}
\arguments{
\item{Mat}{An Matrix}

\item{center}{Whether to center or not: TRUE or FALSE}

\item{scale}{Whether to scale or not: TRUE or FALSE}
}
\value{
Mat The Matrix after processing
}
\description{
the function to center and scale the Bag of Word Matrix
first scale then center, otherwise NaN's can be produced.
}
\keyword{Center}
\keyword{Scale}
