% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM.R
\name{createB}
\alias{createB}
\title{createB}
\usage{
createB(AdjMat, kms_cluster, kmr_cluster)
}
\arguments{
\item{AdjMat}{An Adjacency Matrix}

\item{kms_cluster}{A vector of cluster assignments for the first type of nodes (senders)}

\item{kmr_cluster}{A vector of cluster assignments for the second type of nodes (receivers)}
}
\value{
SbmMat is the SBM matrix
}
\description{
Create SBM matrix
}
\keyword{Matrix}
\keyword{SBM}
