% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeCon.R
\name{laplacian}
\alias{laplacian}
\title{----------------------------------------------------------------------
Helper functions
----------------------------------------------------------------------
laplacian}
\usage{
laplacian(AdjMat, type, regularizer)
}
\arguments{
\item{AdjMat}{An Adjacency Matrix}

\item{type}{Type to normalize to matrix: only by rows, only by columns, or both}

\item{regularizer}{Whether to regularize: TRUE or FALSE}
}
\value{
GraphLap The Graph Laplacian
}
\description{
The function to calculate Graph Laplacian from an Adjacency Matrix
}
\keyword{Graph}
\keyword{Laplacian}
