% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeCon.R
\name{SpeCon}
\alias{SpeCon}
\title{SpeCon}
\usage{
SpeCon(AdjMat, BowMatX, BowMatY, signif_level, weight_h, ncluster_x, ncluster_y,
  niteration)
}
\arguments{
\item{AdjMat}{An Adjacency Matrix}

\item{BowMatX}{A Bag of Word Matrix for the first type of nodes (senders)}

\item{BowMatY}{A Bag of Word Matrix for the second type of nodes (receivers)
The two types of nodes can be the same.}

\item{signif_level}{Significance level that decide the percentage of
non-zero elements of the Call Response Matrix we use. (Thresholding Step)}

\item{weight_h}{The weight h for the covariate-assisted part}

\item{ncluster_x}{Number of clusters for the first type of nodes (senders)}

\item{ncluster_y}{Number of clusters for the second type of nodes (receivers)}

\item{niteration}{Number of iterations in k-means.}
}
\value{
A list of parameters:
svdSimMat Singular values and singular vectors of the similarity matrix.
km_x Clustering results for the first type of nodes (senders)
km_y Clustering results for the second type of nodes (receivers)
}
\description{
SpeCon
}
\keyword{Contextualization}
\keyword{Spectral}
