% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeCon.R
\name{threshold}
\alias{threshold}
\title{threshold
the function to threshold Matrix}
\usage{
threshold(Mat, thresh)
}
\arguments{
\item{Mat}{A matrix}

\item{thresh}{Threshold}
}
\value{
Mat_thresh The matrix after thresholding
}
\description{
threshold
the function to threshold Matrix
}
\keyword{Threshold}
